requirements.txt
cloudpickle==0.6.1
cycler==0.10.0
dask==0.20.2
decorator==4.3.0
fpdf==1.7.2
kiwisolver==1.0.1
matplotlib==3.0.2
networkx==2.2
numpy==1.15.4
opencv-python==3.4.3.18
Pillow==5.3.0
pyparsing==2.3.0
PyQt5==5.11.3
PyQt5-sip==4.19.13
python-dateutil==2.7.5
PyWavelets==1.0.1
scikit-image==0.14.1
scipy==1.1.0
six==1.11.0
toolz==0.9.0
